pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "finadviser"
version = "0.1.0"
description = "Personal financial adviser TUI application"
requires-python = ">=3.9"
dependencies = [
    "textual>=0.85.0",
    "rich>=13.0",
    "pandas>=2.0",
    "anthropic>=0.40.0",
    "pydantic>=2.0",
    "pyyaml>=6.0",
    "click>=8.0",
    "eval_type_backport>=0.2.0; python_version < '3.10'",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
    "textual-dev>=1.0",
]

[project.scripts]
finadviser = "finadviser.cli:main"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
